ff2r_default_abilities: Fix CBS randomly getting Bat weapon (#112) [M]
Without the fix `length` will be equal to 5, thus `GetURandomInt()` will
return a number in the range 0 to 4 inclusive. But 4 is the index of
"plugin_name" (out of range for our situation), so `GetInt()` will
return default value of 0.

So decrease length so it's 4 and thus `GetURandomInt()` will return in
range 0 to 3. `GetKey()` indexing starts from 0.
You get it.